requirements.txt
 
boto3==1.26.137
requests==2.31.0
pandas==2.0.3
pyspark==3.4.0
python-dotenv==1.0.0
pytest==7.4.0
jupyter==1.0.0